added ingestion script, moved utils
